Removed RED3ArchiveManager classes
