+++ WolvenKit.Modkit/Managers/ArchiveManager.cs
@@ -3,7 +3,10 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Text;
 using System.Text.Json;
+using System.Threading.Tasks;
+using CommunityToolkit.Mvvm.ComponentModel;
 using DynamicData;
 using DynamicData.Kernel;
 using WolvenKit.Common;
@@ -11,13 +14,12 @@ using WolvenKit.Common.Model;
 using WolvenKit.Common.Services;
 using WolvenKit.Core.Extensions;
 using WolvenKit.Core.Interfaces;
-using WolvenKit.RED4.Archive;
 using WolvenKit.RED4.Archive.CR2W;
 using WolvenKit.RED4.Types;
 
 namespace WolvenKit.RED4.CR2W.Archive
 {
-    public class ArchiveManager : RED4ArchiveManager
+    public class ArchiveManager : ObservableObject, IArchiveManager
     {
         #region Fields
 
@@ -39,6 +41,18 @@ namespace WolvenKit.RED4.CR2W.Archive
 
         #endregion Fields
 
+        #region Properties
+
+        public RedFileSystemModel? RootNode { get; set; }
+
+        public List<RedFileSystemModel> ModRoots { get; set; } = new();
+
+        public IEnumerable<string>? Extensions { get; set; }
+
+        public bool IsModBrowserActive { get; set; }
+
+        #endregion Properties
+
         #region Constructors
 
         public ArchiveManager(IHashService hashService, Red4ParserService wolvenkitFileService, ILoggerService logger)
@@ -55,31 +69,31 @@ namespace WolvenKit.RED4.CR2W.Archive
 
         #region properties
 
-        public override bool IsManagerLoading
+        public bool IsManagerLoading
         {
             get => _isManagerLoading;
             set => SetProperty(ref _isManagerLoading, value);
         }
 
-        public override bool IsManagerLoaded
+        public bool IsManagerLoaded
         {
             get => _isManagerLoaded;
             set => SetProperty(ref _isManagerLoaded, value);
         }
 
 
-        public override SourceCache<IGameArchive, string> Archives { get; set; } = new(x => x.ArchiveAbsolutePath);
+        public SourceCache<IGameArchive, string> Archives { get; set; } = new(x => x.ArchiveAbsolutePath);
 
         public IObservable<IChangeSet<IGameArchive, string>> ConnectArchives() => Archives.Connect();
 
 
-        public override IObservable<IChangeSet<RedFileSystemModel>> ConnectGameRoot() => _rootCache.Connect();
+        public IObservable<IChangeSet<RedFileSystemModel>> ConnectGameRoot() => _rootCache.Connect();
 
-        public override IObservable<IChangeSet<RedFileSystemModel>> ConnectModRoot() => _modCache.Connect();
+        public IObservable<IChangeSet<RedFileSystemModel>> ConnectModRoot() => _modCache.Connect();
 
-        public override EArchiveType TypeName => EArchiveType.Archive;
+        public EArchiveType TypeName => EArchiveType.Archive;
 
-        public override IGameArchive? ProjectArchive { get; set; }
+        public IGameArchive? ProjectArchive { get; set; }
 
         #endregion properties
 
@@ -150,7 +164,7 @@ namespace WolvenKit.RED4.CR2W.Archive
         /// Loads all archives from a folder
         /// </summary>
         /// <param name="archiveDir"></param>
-        public override void LoadFromFolder(DirectoryInfo archiveDir)
+        public void LoadFromFolder(DirectoryInfo archiveDir)
         {
             if (!archiveDir.Exists)
             {
@@ -181,7 +195,7 @@ namespace WolvenKit.RED4.CR2W.Archive
         /// </summary>
         /// <param name="executable"></param>
         /// <param name="rebuildtree"></param>
-        public override void LoadGameArchives(FileInfo executable, bool rebuildtree = true)
+        public void LoadGameArchives(FileInfo executable, bool rebuildtree = true)
         {
             var di = executable.Directory;
             if (di?.Parent?.Parent is null)
@@ -260,12 +274,52 @@ namespace WolvenKit.RED4.CR2W.Archive
             IsManagerLoaded = true;
         }
 
+        protected void RebuildGameRoot(IHashService hashService)
+        {
+            RootNode = new RedFileSystemModel(TypeName.ToString());
+
+            var allFiles = GetGameArchives()
+                .SelectMany(x => x.Files)
+                .GroupBy(x => x.Key)
+                .Select(x => x.First());
+
+            Parallel.ForEach(allFiles, (file) =>
+            {
+                var path = hashService.Get(file.Key);
+                if (path == null)
+                {
+                    RootNode.Files.Enqueue(file.Value);
+                    return;
+                }
+
+                var lastNode = RootNode;
+
+                var sb = new StringBuilder();
+                for (var i = 0; i < path.Length; i++)
+                {
+                    if (path[i] == '\\')
+                    {
+                        var str = sb.ToString();
+
+                        if (!lastNode.Directories.TryGetValue(str, out var tmpNode))
+                        {
+                            tmpNode = new RedFileSystemModel(str);
+                            lastNode.Directories.TryAdd(str, tmpNode);
+                        }
+                        lastNode = tmpNode;
+                    }
+                    sb.Append(path[i]);
+                }
+                lastNode.Files.Enqueue(file.Value);
+            });
+        }
+
         /// <summary>
         /// Load a single bundle
         /// </summary>
         /// <param name="path"></param>
         /// <param name="source"></param>
-        public override void LoadArchive(string path, EArchiveSource source = EArchiveSource.Unknown)
+        public void LoadArchive(string path, EArchiveSource source = EArchiveSource.Unknown)
         {
             if (Archives.Lookup(path).HasValue)
             {
@@ -291,7 +345,7 @@ namespace WolvenKit.RED4.CR2W.Archive
         /// file to process
         /// </param>
         /// <param name="analyzeFiles"></param>
-        public override void LoadModArchive(string filename, bool analyzeFiles = true)
+        public void LoadModArchive(string filename, bool analyzeFiles = true)
         {
             if (Archives.Lookup(filename).HasValue)
             {
@@ -343,7 +397,7 @@ namespace WolvenKit.RED4.CR2W.Archive
         /// <summary>
         /// Loads bundles from specified mods and dlc folder
         /// </summary>
-        public override void LoadModsArchives(FileInfo executable, bool analyzeFiles = true)
+        public void LoadModsArchives(FileInfo executable, bool analyzeFiles = true)
         {
             var di = executable.Directory;
             if (di?.Parent?.Parent is null)
@@ -484,7 +538,7 @@ namespace WolvenKit.RED4.CR2W.Archive
             IsManagerLoaded = true;
         }
 
-        public override void LoadAdditionalModArchives(string archiveBasePath, bool analyzeFiles = true)
+        public void LoadAdditionalModArchives(string archiveBasePath, bool analyzeFiles = true)
         {
             if (!Directory.Exists(archiveBasePath))
             {
@@ -534,13 +588,52 @@ namespace WolvenKit.RED4.CR2W.Archive
             IsManagerLoaded = true;
         }
 
+        protected void RebuildModRoot()
+        {
+            ModRoots.Clear();
+
+            foreach (var archive in GetModArchives())
+            {
+                ArgumentNullException.ThrowIfNull(archive.ArchiveRelativePath);
+
+                var modroot = new RedFileSystemModel(archive.ArchiveRelativePath);
+
+                // loop through all files
+                //Parallel.ForEach(archive.Files, item =>
+                foreach (var item in archive.Files)
+                {
+                    var currentNode = modroot;
+                    var model = item.Value;
+                    var parts = model.Name.Split('\\');
+
+                    // loop through path
+                    var fullpath = "";
+                    for (var i = 0; i < parts.Length - 1; i++)
+                    {
+                        var part = parts[i];
+                        fullpath += part + Path.DirectorySeparatorChar;
+
+                        var dirPath = fullpath.TrimEnd(Path.DirectorySeparatorChar);
+                        var x = currentNode.Directories.GetOrAdd(dirPath, new RedFileSystemModel(dirPath));
+                        currentNode = x;
+                    }
+
+                    // add file to the last directory in path
+                    currentNode.Files.Enqueue(item.Value);
+                }
+                //);
+
+                ModRoots.Add(modroot);
+            }
+        }
+
         #endregion
 
         /// <summary>
         /// Get files grouped by extension in all archives of current search scope
         /// </summary>
         /// <returns></returns>
-        public override Dictionary<string, IEnumerable<IGameFile>> GetGroupedFiles()
+        public Dictionary<string, IEnumerable<IGameFile>> GetGroupedFiles()
         {
             if (IsModBrowserActive)
             {
@@ -566,10 +659,10 @@ namespace WolvenKit.RED4.CR2W.Archive
         public bool ContainsFile(ulong hash) => Lookup(hash).HasValue;
 
         /// <inheritdoc />
-        public override Optional<IGameFile> Lookup(ulong hash) => Lookup(hash, ArchiveManagerScope.Everywhere);
+        public Optional<IGameFile> Lookup(ulong hash) => Lookup(hash, ArchiveManagerScope.Everywhere);
 
         /// <inheritdoc />
-        public override Optional<IGameFile> Lookup(ulong hash, ArchiveManagerScope searchScope)
+        public Optional<IGameFile> Lookup(ulong hash, ArchiveManagerScope searchScope)
         {
             if (searchScope is ArchiveManagerScope.Mods or ArchiveManagerScope.Everywhere)
             {
@@ -607,7 +700,7 @@ namespace WolvenKit.RED4.CR2W.Archive
             return Optional<IGameFile>.None;
         }
 
-        public override IGameFile? GetGameFile(ResourcePath path, bool includeMods = true, bool includeProject = true)
+        public IGameFile? GetGameFile(ResourcePath path, bool includeMods = true, bool includeProject = true)
         {
             // check if the file is in the project archive
             if (includeProject && ProjectArchive != null && ProjectArchive.Files.TryGetValue(path, out var projectFile))
@@ -647,7 +740,7 @@ namespace WolvenKit.RED4.CR2W.Archive
             return null;
         }
 
-        public override CR2WFile? GetCR2WFile(ResourcePath path, bool includeMods = true, bool includeProject = true)
+        public CR2WFile? GetCR2WFile(ResourcePath path, bool includeMods = true, bool includeProject = true)
         {
             var gameFile = GetGameFile(path, includeMods, includeProject);
 
@@ -665,6 +758,11 @@ namespace WolvenKit.RED4.CR2W.Archive
 
             return null;
         }
+
+        public IEnumerable<IGameArchive> GetModArchives() => Archives.Items.Where(x => x.Source is EArchiveSource.Mod);
+        public IEnumerable<IGameArchive> GetBaseArchives() => Archives.Items.Where(x => x.Source is EArchiveSource.Base);
+        public IEnumerable<IGameArchive> GetEp1Archives() => Archives.Items.Where(x => x.Source is EArchiveSource.EP1);
+        public IEnumerable<IGameArchive> GetGameArchives() => Archives.Items.Where(x => x.Source is EArchiveSource.EP1 or EArchiveSource.Base);
 
         #endregion methods
     }
