--- WolvenKit.Modkit/Managers/RED4ArchiveManager.cs
@@ -1,15 +1,0 @@
-using WolvenKit.Common.Model;
-
-namespace WolvenKit.Common
-{
-    /// <summary>
-    /// Abstract implementation for cyberpunk 2077 archive managers
-    /// e.g. archive
-    /// </summary>
-    public abstract class RED4ArchiveManager : WolvenKitArchiveManager
-    {
-        public RED4ArchiveManager()
-        {
-        }
-    }
-}
